Move src/pytorch_lightning/lite to src/lightning_lite (#14735)
