+++ examples/convert_from_pt_to_pl/image_classifier_2_lite.py
@@ -38,10 +38,10 @@ import torchvision.transforms as T
 from torch.optim.lr_scheduler import StepLR
 from torchmetrics.classification import Accuracy
 
+from lightning_lite.lite import LightningLite  # import LightningLite
 from pytorch_lightning import seed_everything
 from pytorch_lightning.demos.boring_classes import Net
 from pytorch_lightning.demos.mnist_datamodule import MNIST
-from pytorch_lightning.lite import LightningLite  # import LightningLite
 
 DATASETS_PATH = path.join(path.dirname(__file__), "..", "..", "Datasets")
