+++ examples/convert_from_pt_to_pl/image_classifier_3_lite_to_lightning_module.py
@@ -34,10 +34,10 @@ import torchvision.transforms as T
 from torch.optim.lr_scheduler import StepLR
 from torchmetrics import Accuracy
 
+from lightning_lite.lite import LightningLite
 from pytorch_lightning import seed_everything
 from pytorch_lightning.demos.boring_classes import Net
 from pytorch_lightning.demos.mnist_datamodule import MNIST
-from pytorch_lightning.lite import LightningLite
 
 DATASETS_PATH = path.join(path.dirname(__file__), "..", "..", "Datasets")
