+++ examples/pl_loops/mnist_lite.py
@@ -22,10 +22,10 @@ import torchvision.transforms as T
 from torch.optim.lr_scheduler import StepLR
 from torchmetrics import Accuracy
 
+from lightning_lite.lite import LightningLite
 from pytorch_lightning import seed_everything
 from pytorch_lightning.demos.boring_classes import Net
 from pytorch_lightning.demos.mnist_datamodule import MNIST
-from pytorch_lightning.lite import LightningLite
 from pytorch_lightning.loops import Loop
 
 DATASETS_PATH = path.join(path.dirname(__file__), "..", "..", "Datasets")
