+++ tests/tests_lite/test_lite.py
@@ -20,17 +20,17 @@ import pytest
 import torch
 import torch.distributed
 import torch.nn.functional
+from tests_lite.helpers.runif import RunIf
 from torch import nn
 from torch.utils.data import DataLoader, DistributedSampler, Sampler
 
+from lightning_lite.lite import LightningLite
 from lightning_lite.plugins import Precision
 from lightning_lite.strategies import DeepSpeedStrategy, Strategy
 from lightning_lite.utilities import _StrategyType
 from lightning_lite.utilities.exceptions import MisconfigurationException
 from lightning_lite.utilities.seed import pl_worker_init_function
-from pytorch_lightning.lite import LightningLite
-from pytorch_lightning.lite.wrappers import _LiteDataLoader, _LiteModule, _LiteOptimizer
-from tests_pytorch.helpers.runif import RunIf
+from lightning_lite.wrappers import _LiteDataLoader, _LiteModule, _LiteOptimizer
 
 
 class EmptyLite(LightningLite):
@@ -165,7 +165,7 @@ def test_setup_dataloaders_return_type():
     assert lite_dataloader1.dataset is dataset1
 
 
-@mock.patch("pytorch_lightning.lite.lite._replace_dunder_methods")
+@mock.patch("lightning_lite.lite._replace_dunder_methods")
 def test_setup_dataloaders_captures_dataloader_arguments(ctx_manager):
     """Test that Lite intercepts the DataLoader constructor arguments with a context manager in its run method."""
 
@@ -210,7 +210,7 @@ def test_setup_dataloaders_twice_fails():
 
 
 @mock.patch(
-    "pytorch_lightning.lite.lite.LightningLite.device",
+    "lightning_lite.lite.LightningLite.device",
     new_callable=PropertyMock,
     return_value=torch.device("cuda", 1),
 )
