+++ tests/tests_lite/test_wrappers.py
@@ -15,12 +15,12 @@ from unittest.mock import Mock
 
 import pytest
 import torch
+from tests_lite.helpers.runif import RunIf
 from torch.utils.data.dataloader import DataLoader
 
+from lightning_lite.lite import LightningLite
 from lightning_lite.utilities.device_dtype_mixin import _DeviceDtypeModuleMixin
-from pytorch_lightning.lite import LightningLite
-from pytorch_lightning.lite.wrappers import _LiteDataLoader, _LiteModule, _LiteOptimizer
-from tests_pytorch.helpers.runif import RunIf
+from lightning_lite.wrappers import _LiteDataLoader, _LiteModule, _LiteOptimizer
 
 
 class EmptyLite(LightningLite):
